ajout Serie et corrections

+++ src/main/java/com/java/planta_criteria/commune/CommuneRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface CommuneRepository extends JpaRepository<Commune, Long> {
+public interface CommuneRepository extends JpaRepository<Commune, Integer> {
 
     List<Commune> findTop10ByNomContainingIgnoreCase(String nom);
 }
