+++ src/main/java/com/java/planta_criteria/critere/CritereNotFoundException.java
@@ -2,7 +2,7 @@ package com.java.planta_criteria.critere;
 
 public class CritereNotFoundException extends RuntimeException {
 
-    public CritereNotFoundException(Long id) {
+    public CritereNotFoundException(Integer id) {
         super("Critère introuvable : " + id);
     }
 }
