+++ src/main/java/com/java/planta_criteria/image/ImageNotFoundException.java
@@ -2,7 +2,7 @@ package com.java.planta_criteria.image;
 
 public class ImageNotFoundException extends RuntimeException {
 
-    public ImageNotFoundException(Long id) {
+    public ImageNotFoundException(Integer id) {
         super("Image introuvable : " + id);
     }
 }
