+++ src/main/java/com/java/planta_criteria/image/ImageRepository.java
@@ -2,5 +2,5 @@ package com.java.planta_criteria.image;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ImageRepository extends JpaRepository<Image, Long> {
+public interface ImageRepository extends JpaRepository<Image, Integer> {
 }
