+++ src/main/java/com/java/planta_criteria/observation/ObservationNotFoundException.java
@@ -2,7 +2,7 @@ package com.java.planta_criteria.observation;
 
 public class ObservationNotFoundException extends RuntimeException {
 
-    public ObservationNotFoundException(Long id) {
+    public ObservationNotFoundException(Integer id) {
         super("Observation introuvable : " + id);
     }
 }
