+++ src/main/java/com/java/planta_criteria/observation/ObservationRepository.java
@@ -2,5 +2,5 @@ package com.java.planta_criteria.observation;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ObservationRepository extends JpaRepository<Observation, Long> {
+public interface ObservationRepository extends JpaRepository<Observation, Integer> {
 }
