+++ src/main/java/com/java/planta_criteria/observation_critere/ObservationCritereNotFoundException.java
@@ -3,7 +3,7 @@ package com.java.planta_criteria.observation_critere;
 public class ObservationCritereNotFoundException
         extends RuntimeException {
 
-    public ObservationCritereNotFoundException(Long id) {
+    public ObservationCritereNotFoundException(Integer id) {
         super("ObservationCritere introuvable : " + id);
     }
 }
