+++ src/main/java/com/java/planta_criteria/observation_critere/ObservationCritereRepository.java
@@ -3,5 +3,5 @@ package com.java.planta_criteria.observation_critere;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ObservationCritereRepository
-        extends JpaRepository<ObservationCritere, Long> {
+        extends JpaRepository<ObservationCritere, Integer> {
 }
