+++ src/main/java/com/java/planta_criteria/observation_critere/ObservationCritereService.java
@@ -29,7 +29,7 @@ public class ObservationCritereService {
             .toList();
     }
 
-    public ObservationCritereDto findById(Long id) {
+    public ObservationCritereDto findById(Integer id) {
 
         ObservationCritere observationCritere =
             repository.findById(id)
