+++ src/main/java/com/java/planta_criteria/source/SourceNotFoundException.java
@@ -2,7 +2,7 @@ package com.java.planta_criteria.source;
 
 public class SourceNotFoundException extends RuntimeException {
 
-    public SourceNotFoundException(Long id) {
+    public SourceNotFoundException(Integer id) {
         super("Source introuvable : " + id);
     }
 }
