+++ src/main/java/com/java/planta_criteria/source/SourceRepository.java
@@ -2,5 +2,5 @@ package com.java.planta_criteria.source;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SourceRepository extends JpaRepository<Source, Long> {
+public interface SourceRepository extends JpaRepository<Source, Integer> {
 }
