+++ src/main/java/com/java/planta_criteria/statut/StatutRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface StatutRepository extends JpaRepository<Statut, Long> {
+public interface StatutRepository extends JpaRepository<Statut, Integer> {
 
     Optional<Statut> findByLibelle(String libelle);
 }
