+++ src/main/java/com/java/planta_criteria/taxref/TaxrefNotFoundException.java
@@ -3,7 +3,7 @@ package com.java.planta_criteria.taxref;
 
 public class TaxrefNotFoundException extends RuntimeException {
 
-    public TaxrefNotFoundException(Long id) {
+    public TaxrefNotFoundException(Integer id) {
         super("Taxon introuvable : " + id);
     }
 }
